moved fun to clear DB
