+++ app/src/main/java/com/mackenzie/waifuviewer/ui/selector/SelectorViewModel.kt
@@ -8,10 +8,13 @@ import com.mackenzie.waifuviewer.domain.WaifuBestItem
 import com.mackenzie.waifuviewer.domain.WaifuPicItem
 import com.mackenzie.waifuviewer.domain.im.WaifuImItem
 import com.mackenzie.waifuviewer.domain.im.WaifuImTagList
+import com.mackenzie.waifuviewer.usecases.best.ClearWaifuBestUseCase
 import com.mackenzie.waifuviewer.usecases.best.RequestOnlyWaifuBestUseCase
+import com.mackenzie.waifuviewer.usecases.im.ClearWaifuImUseCase
 import com.mackenzie.waifuviewer.usecases.im.GetWaifuImTagsUseCase
 import com.mackenzie.waifuviewer.usecases.im.RequestOnlyWaifuImUseCase
 import com.mackenzie.waifuviewer.usecases.im.RequestWaifuImTagsUseCase
+import com.mackenzie.waifuviewer.usecases.pics.ClearWaifuPicUseCase
 import com.mackenzie.waifuviewer.usecases.pics.RequestOnlyWaifuPicUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -28,7 +31,10 @@ class SelectorViewModel @Inject constructor(
     private val requestOnlyImWaifu: RequestOnlyWaifuImUseCase,
     private val requestOnlyPicWaifu: RequestOnlyWaifuPicUseCase,
     private val requestOnlyWaifuBestUseCase: RequestOnlyWaifuBestUseCase,
-    private val requestWaifuImTagsUseCase: RequestWaifuImTagsUseCase
+    private val requestWaifuImTagsUseCase: RequestWaifuImTagsUseCase,
+    private val clearWaifuImUseCase: ClearWaifuImUseCase,
+    private val clearWaifuPicUseCase: ClearWaifuPicUseCase,
+    private val clearWaifuBestUseCase: ClearWaifuBestUseCase
 ): ViewModel() {
 
     private val _state = MutableStateFlow(UiState())
@@ -90,6 +96,27 @@ class SelectorViewModel @Inject constructor(
             ServerType.ENHANCED -> loadPicWaifu()
             ServerType.NEKOS -> loadNekoWaifu()
             else -> {}
+        }
+    }
+
+    fun onClearImDatabase() {
+        viewModelScope.launch {
+            val error = clearWaifuImUseCase()
+            _state.update { it.copy(error = error) }
+        }
+    }
+
+    fun onClearPicsDatabase() {
+        viewModelScope.launch {
+            val error = clearWaifuPicUseCase()
+            _state.update { _state.value.copy(error = error) }
+        }
+    }
+
+    fun onClearBestDatabase() {
+        viewModelScope.launch {
+            val error = clearWaifuBestUseCase()
+            _state.update { _state.value.copy(error = error) }
         }
     }
